+++ crates/apheleia-core/src/types.rs
@@ -4,8 +4,8 @@ use std::cmp;
 /// A simple struct that stores x, y: u16
 #[derive(Clone, Copy, fmt::Debug, cmp::PartialEq)]
 pub struct Vec2 {
-    pub x: u16,
-    pub y: u16,
+    pub x: u32,
+    pub y: u32,
 }
 impl Vec2 {
     /// Returns a [`Vec2`] where x = y = 0
